requirements.txt
academictorrents==2.3.3
albumentations==2.0.8
caffe2==0.8.1
click==8.1.7
diffusers==0.29.1
docutils==0.22.4
einops==0.8.1
fairscale==0.4.13
ftfy==6.3.1
joblib==1.4.2
kneed==0.8.5
kornia==0.7.3
lmdb==1.5.1
matplotlib==3.7.5
natsort==8.4.0
networkx==3.1
numpy==1.24.4
omegaconf==2.3.0
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
packaging==25.0
pandas==2.0.3
Pillow==12.0.0
protobuf==6.33.2
pycocotools==2.0.7
pynvml==13.0.1
pytorch_lightning==1.5.1
PyYAML==6.0.2
PyYAML==6.0.3
qwen_vl_utils==0.0.14
rasterio==1.3.11
regex==2024.9.11
scikit_learn==1.3.2
scipy==1.16.3
seaborn==0.13.2
setuptools==60.2.0
setuptools==75.3.0
skimage==0.0
Sphinx==9.0.4
sphinx_rtd_theme==3.0.2
tabulate==0.9.0
timm==1.0.22
torch==2.3.1+cu118
torch==2.4.1+cu121
torch_geometric==2.7.0
torchvision==0.18.1+cu118
tqdm==4.65.0
transformers==4.46.3
triton==3.5.1
xformers==0.0.33.post2
xmlrunner==1.7.7
